+++ src/effects/VST3/VST3Effect.h
@@ -15,28 +15,11 @@
 
 #include <wx/wx.h>
 
-#include <pluginterfaces/gui/iplugview.h>
 #include <public.sdk/source/vst/hosting/module.h>
 
 #include "../StatefulPerTrackEffect.h"
 
 class VST3Instance;
-class NumericTextCtrl;
-
-namespace Steinberg
-{
-   namespace Vst
-   {
-      class IParameterChanges;
-      class IComponent;
-      class IEditController;
-      class IConnectionPoint;
-   }
-}
-
-class ParameterChangesProvider;
-class VST3ParametersWindow;
-class VST3Wrapper;
 
 /**
  * \brief Objects of this class connect Audacity with VST3 effects
@@ -46,6 +29,8 @@ class VST3Effect final : public PerTrackEffect
    // Keep strong reference to a module; this because it has to be destroyed in the destructor of this class,
    // otherwise the destruction of mEditController and mEffectComponent would trigger a memory fault.
    std::shared_ptr<VST3::Hosting::Module> mModule;
+
+   wxWindow* mParent{nullptr};
 
    VST3::Hosting::ClassInfo mEffectClassInfo;
 
@@ -84,26 +69,24 @@ public:
    bool IsDefault() const override;
    RealtimeSince RealtimeSupport() const override;
    bool SupportsAutomation() const override;
-   bool SaveSettings(
-      const EffectSettings &settings, CommandParameters & parms) const override;
-   bool LoadSettings(
-      const CommandParameters & parms, EffectSettings &settings) const override;
-   bool LoadUserPreset(
-      const RegistryPath & name, EffectSettings &settings) const override;
-   bool SaveUserPreset(
-      const RegistryPath & name, const EffectSettings &settings) const override;
+   ///Saves only changes recorded during session, not the complete state(see SaveUserPreset)
+   bool SaveSettings(const EffectSettings &settings, CommandParameters & parms) const override;
+   ///Applies previousely saved parameter changes
+   bool LoadSettings( const CommandParameters & parms, EffectSettings &settings) const override;
+   ///Attempts to restore previousely saved state to the currently edited(!) effect
+   bool LoadUserPreset(const RegistryPath & name, EffectSettings &settings) const override;
+   ///Attempts to store the complete state of the currently edited(!) effect
+   bool SaveUserPreset(const RegistryPath & name, const EffectSettings &settings) const override;
    RegistryPaths GetFactoryPresets() const override;
    bool LoadFactoryPreset(int id, EffectSettings &settings) const override;
 
    int ShowClientInterface(wxWindow &parent, wxDialog &dialog,
       EffectUIValidator *pValidator, bool forceModal) override;
+   bool CloseUI() override;
    std::shared_ptr<EffectInstance> MakeInstance() const override;
-   bool IsGraphicalUI() override;
    std::unique_ptr<EffectUIValidator> PopulateUI(
       ShuttleGui &S, EffectInstance &instance, EffectSettingsAccess &access)
    override;
-   bool ValidateUI(EffectSettings &) override;
-   bool CloseUI() override;
    bool CanExportPresets() override;
    void ExportPresets(const EffectSettings &settings) const override;
    void ImportPresets(EffectSettings &settings) override;
@@ -111,9 +94,7 @@ public:
    void ShowOptions() override;
 
    EffectSettings MakeSettings() const override;
-
-   bool TransferDataToWindow(const EffectSettings& settings) override;
-
+   
 private:
 
    bool LoadPreset(const wxString& path, EffectSettings& settings);
